webcraper: Read firefox/geckodriver paths from config file
To prevent git related conflicts with the "FIREFOX_BINARY" and "GECKODRIVER"
variables, which will probably be different between each computer, I've
moved these to a config file.

A sample config will be added to the git repository, which will be copied
to 'config.json', and edited with the correct paths for 'firefox' and
'geckodriver'.